Set up a function for bootstrapping
Bootstrap function reads in a BED file to determine sampling regions,
determines the proper region for each SNP, and then creates a new SFS
for each iteration of the bootstrap process and returns them as a list.
Default separation is by chromosome, and it is possible to write to
disk rather than store as a list.